+++ ChangeLog.md
@@ -3,7 +3,7 @@
 ## Linked Data 0.0.1.0
 - Basic datatypes and their arbitrary instances defined (both heavily influenced
   by rdf4h).
-- Turtle parser and N-Triples parser implemented.
+- Turtle parser and serialiser and N-Triples parser and serialiser implemented.
 - Isomorphism algorithm from Aidan Hogan's paper implemented.
 - Basic query functionality implemented.
 - Test files from W3C added.